refactor(interpeter): Split out stack
