+++ liquid-interpreter/src/context.rs
@@ -1,12 +1,12 @@
-use std::borrow;
 use std::collections::HashMap;
 use std::sync;
 
 use error::{Error, Result};
 use itertools;
-use value::{Object, PathRef, Scalar, Value};
+use value::Value;
 
 use super::Expression;
+use super::Stack;
 use super::ValueStore;
 use super::PluginRegistry;
 use super::{BoxedValueFilter, FilterValue};
@@ -122,171 +122,6 @@ impl IfChangedState {
         self.last_rendered = Some(rendered.to_owned());
 
         has_changed
-    }
-}
-
-/// Stack of variables.
-#[derive(Debug, Clone)]
-pub struct Stack<'g> {
-    globals: Option<&'g ValueStore>,
-    stack: Vec<Object>,
-    // State of variables created through increment or decrement tags.
-    indexes: Object,
-}
-
-impl<'g> Stack<'g> {
-    /// Create an empty stack
-    pub fn empty() -> Self {
-        Self {
-            globals: None,
-            indexes: Object::new(),
-            // Mutable frame for globals.
-            stack: vec![Object::new()],
-        }
-    }
-
-    /// Create a stack initialized with read-only `ValueStore`.
-    pub fn with_globals(globals: &'g ValueStore) -> Self {
-        let mut stack = Self::empty();
-        stack.globals = Some(globals);
-        stack
-    }
-
-    /// Creates a new variable scope chained to a parent scope.
-    fn push_frame(&mut self) {
-        self.stack.push(Object::new());
-    }
-
-    /// Removes the topmost stack frame from the local variable stack.
-    ///
-    /// # Panics
-    ///
-    /// This method will panic if popping the topmost frame results in an
-    /// empty stack. Given that a context is created with a top-level stack
-    /// frame already in place, emptying the stack should never happen in a
-    /// well-formed program.
-    fn pop_frame(&mut self) {
-        if self.stack.pop().is_none() {
-            panic!("Unbalanced push/pop, leaving the stack empty.")
-        };
-    }
-
-    /// Recursively index into the stack.
-    pub fn try_get(&self, path: PathRef) -> Option<&Value> {
-        let frame = self.find_path_frame(path)?;
-
-        frame.try_get_variable(path)
-    }
-
-    /// Recursively index into the stack.
-    pub fn get(&self, path: PathRef) -> Result<&Value> {
-        let frame = self.find_path_frame(path).ok_or_else(|| {
-            let key = path
-                .iter()
-                .next()
-                .cloned()
-                .unwrap_or_else(|| Scalar::new("nil"));
-            let globals = itertools::join(self.globals().iter(), ", ");
-            Error::with_msg("Unknown variable")
-                .context("requested variable", key.to_str().into_owned())
-                .context("available variables", globals)
-        })?;
-
-        frame.get_variable(path)
-    }
-
-    fn globals(&self) -> Vec<&str> {
-        let mut globals = self.globals.map(|g| g.roots()).unwrap_or_default();
-        for frame in self.stack.iter() {
-            globals.extend(frame.roots());
-        }
-        globals.sort();
-        globals.dedup();
-        globals
-    }
-
-    fn find_path_frame<'a>(&'a self, path: PathRef) -> Option<&'a ValueStore> {
-        let key = path.iter().next()?;
-        let key = key.to_str();
-        self.find_frame(key.as_ref())
-    }
-
-    fn find_frame<'a>(&'a self, name: &str) -> Option<&'a ValueStore> {
-        for frame in self.stack.iter().rev() {
-            if frame.contains_root(name) {
-                return Some(frame);
-            }
-        }
-
-        if self
-            .globals
-            .map(|g| g.contains_root(name))
-            .unwrap_or(false)
-        {
-            return self.globals;
-        }
-
-        if self.indexes.contains_root(name) {
-            return Some(&self.indexes);
-        }
-
-        None
-    }
-
-    /// Used by increment and decrement tags
-    pub fn set_index<S>(&mut self, name: S, val: Value) -> Option<Value>
-    where
-        S: Into<borrow::Cow<'static, str>>,
-    {
-        self.indexes.insert(name.into(), val)
-    }
-
-    /// Used by increment and decrement tags
-    pub fn get_index<'a>(&'a self, name: &str) -> Option<&'a Value> {
-        self.indexes.get(name)
-    }
-
-    /// Sets a value in the global context.
-    pub fn set_global<S>(&mut self, name: S, val: Value) -> Option<Value>
-    where
-        S: Into<borrow::Cow<'static, str>>,
-    {
-        self.global_frame().insert(name.into(), val)
-    }
-
-    /// Sets a value to the rendering context.
-    /// Note that it needs to be wrapped in a liquid::Value.
-    ///
-    /// # Panics
-    ///
-    /// Panics if there is no frame on the local values stack. Context
-    /// instances are created with a top-level stack frame in place, so
-    /// this should never happen in a well-formed program.
-    pub fn set<S>(&mut self, name: S, val: Value) -> Option<Value>
-    where
-        S: Into<borrow::Cow<'static, str>>,
-    {
-        self.current_frame().insert(name.into(), val)
-    }
-
-    fn current_frame(&mut self) -> &mut Object {
-        match self.stack.last_mut() {
-            Some(frame) => frame,
-            None => panic!("Global frame removed."),
-        }
-    }
-
-    fn global_frame(&mut self) -> &mut Object {
-        match self.stack.first_mut() {
-            Some(frame) => frame,
-            None => panic!("Global frame removed."),
-        }
-    }
-}
-
-impl<'g> Default for Stack<'g> {
-    fn default() -> Self {
-        Self::empty()
     }
 }
 
@@ -430,32 +265,6 @@ mod test {
     use super::*;
 
     use value::Scalar;
-
-    #[test]
-    fn stack_find_frame() {
-        let mut ctx = Context::new();
-        ctx.stack_mut().set_global("number", Value::scalar(42f64));
-        assert!(ctx.stack().find_frame("number").is_some(),);
-    }
-
-    #[test]
-    fn stack_find_frame_failure() {
-        let mut ctx = Context::new();
-        let mut post = Object::new();
-        post.insert("number".into(), Value::scalar(42f64));
-        ctx.stack_mut().set_global("post", Value::Object(post));
-        assert!(ctx.stack().find_frame("post.number").is_none());
-    }
-
-    #[test]
-    fn stack_get() {
-        let mut ctx = Context::new();
-        let mut post = Object::new();
-        post.insert("number".into(), Value::scalar(42f64));
-        ctx.stack_mut().set_global("post", Value::Object(post));
-        let indexes = [Scalar::new("post"), Scalar::new("number")];
-        assert_eq!(ctx.stack().get(&indexes).unwrap(), &Value::scalar(42f64));
-    }
 
     #[test]
     fn scoped_variables() {
