+++ liquid-interpreter/src/lib.rs
@@ -26,6 +26,7 @@ mod filter_chain;
 mod store;
 mod registry;
 mod renderable;
+mod stack;
 mod template;
 mod text;
 mod variable;
@@ -35,6 +36,7 @@ pub use self::expression::*;
 pub use self::filter::*;
 pub use self::filter_chain::*;
 pub use self::store::*;
+pub use self::stack::*;
 pub use self::registry::*;
 pub use self::renderable::*;
 pub use self::template::*;
